fix: Working pagination info

+++ microservices/authentication/src/main/scala/shared/responses/PaginatedResponse.scala
@@ -7,8 +7,8 @@ case class PaginatedResponse[A](
   
 case class Meta (
     currentPage: Int,
-    lastPage: Int,
+    lastPage: Long,
     from: Int,
-    to: Int,
+    to: Long,
     total: Long
 )
